server/requirements.txt
click==7.1.1
fastapi==0.54.1
h11==0.9.0
httptools==0.1.1
passlib==1.7.2
psycopg2-binary==2.8.5
pydantic==1.4
PyJWT==1.7.1
SQLAlchemy==1.3.16
starlette==0.13.2
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
